+++ client/src/pages/dashboard/index.jsx
@@ -175,7 +175,29 @@ const DashboardDefault = () => {
 
                 {/* row 2 */}
                 
-                <PlotComponent data={[metrics["cosmos.system.cpu.0"], metrics["cosmos.system.ram"]]}/>
+                <PlotComponent title={'Resources'} data={[metrics["cosmos.system.cpu.0"], metrics["cosmos.system.ram"]]}/>
+               
+                <Grid item xs={12} md={5} lg={4}>
+                    <Grid container alignItems="center" justifyContent="space-between">
+                        <Grid item>
+                            <Typography variant="h5">Income Overview</Typography>
+                        </Grid>
+                        <Grid item />
+                    </Grid>
+                    <MainCard sx={{ mt: 2 }} content={false}>
+                        <Box sx={{ p: 3, pb: 0 }}>
+                            <Stack spacing={2}>
+                                <Typography variant="h6" color="textSecondary">
+                                    This Week Statistics
+                                </Typography>
+                                <Typography variant="h3">$7,650</Typography>
+                            </Stack>
+                        </Box>
+                        <MonthlyBarChart />
+                    </MainCard>
+                </Grid>
+               
+                <PlotComponent title={'Network'} data={[metrics["cosmos.system.netTx"], metrics["cosmos.system.netRx"]]}/>
                 
                 <Grid item xs={12} md={5} lg={4}>
                     <Grid container alignItems="center" justifyContent="space-between">
@@ -209,6 +231,7 @@ const DashboardDefault = () => {
                         <OrdersTable />
                     </MainCard>
                 </Grid>
+
                 <Grid item xs={12} md={5} lg={4}>
                     <Grid container alignItems="center" justifyContent="space-between">
                         <Grid item>